Extract a loadAll helper (#9027)
