+++ src/dataconnect/fileUtils.ts
@@ -1,133 +1,8 @@
 import * as fs from "fs-extra";
 import * as path from "path";
-import * as clc from "colorette";
-import { glob } from "glob";
 
-import { FirebaseError } from "../error";
-import {
-  ConnectorYaml,
-  DataConnectYaml,
-  File,
-  Platform,
-  ServiceInfo,
-  SupportedFrameworks,
-} from "./types";
-import { readFileFromDirectory, wrappedSafeLoad } from "../utils";
-import { Config } from "../config";
-import { DataConnectMultiple } from "../firebaseConfig";
-import { load } from "./load";
+import { Platform, SupportedFrameworks } from "./types";
 import { PackageJSON } from "../frameworks/compose/discover/runtime/node";
-
-export function readFirebaseJson(config?: Config): DataConnectMultiple {
-  if (!config?.has("dataconnect")) {
-    return [];
-  }
-  const validator = (cfg: any) => {
-    if (!cfg["source"]) {
-      throw new FirebaseError("Invalid firebase.json: DataConnect requires `source`");
-    }
-    return {
-      source: cfg["source"],
-    };
-  };
-  const configs = config.get("dataconnect");
-  if (typeof configs === "object" && !Array.isArray(configs)) {
-    return [validator(configs)];
-  } else if (Array.isArray(configs)) {
-    return configs.map(validator);
-  } else {
-    throw new FirebaseError(
-      "Invalid firebase.json: dataconnect should be of the form { source: string }",
-    );
-  }
-}
-
-export async function readDataConnectYaml(sourceDirectory: string): Promise<DataConnectYaml> {
-  const file = await readFileFromDirectory(sourceDirectory, "dataconnect.yaml");
-  const dataconnectYaml = await wrappedSafeLoad(file.source);
-  return validateDataConnectYaml(dataconnectYaml);
-}
-
-function validateDataConnectYaml(unvalidated: any): DataConnectYaml {
-  // TODO: Use json schema for validation here!
-  if (!unvalidated["location"]) {
-    throw new FirebaseError("Missing required field 'location' in dataconnect.yaml");
-  }
-  return unvalidated as DataConnectYaml;
-}
-
-export async function readConnectorYaml(sourceDirectory: string): Promise<ConnectorYaml> {
-  const file = await readFileFromDirectory(sourceDirectory, "connector.yaml");
-  const connectorYaml = await wrappedSafeLoad(file.source);
-  return validateConnectorYaml(connectorYaml);
-}
-
-function validateConnectorYaml(unvalidated: any): ConnectorYaml {
-  // TODO: Add validation
-  return unvalidated as ConnectorYaml;
-}
-
-export async function readGQLFiles(sourceDir: string): Promise<File[]> {
-  if (!fs.existsSync(sourceDir)) {
-    return [];
-  }
-  const files = await glob("**/*.{gql,graphql}", { cwd: sourceDir, absolute: true, nodir: true });
-  return files.map((f) => toFile(sourceDir, f));
-}
-
-function toFile(sourceDir: string, fullPath: string): File {
-  const relPath = path.relative(sourceDir, fullPath);
-  if (!fs.existsSync(fullPath)) {
-    throw new FirebaseError(`file ${fullPath} not found`);
-  }
-  const content = fs.readFileSync(fullPath).toString();
-  return {
-    path: relPath,
-    content,
-  };
-}
-
-// pickService reads firebase.json and returns all services with a given serviceId.
-// If serviceID is not provided and there is a single service, return that.
-export async function pickService(
-  projectId: string,
-  config: Config,
-  serviceId?: string,
-): Promise<ServiceInfo> {
-  const serviceCfgs = readFirebaseJson(config);
-  let serviceInfo: ServiceInfo;
-  if (serviceCfgs.length === 0) {
-    throw new FirebaseError(
-      "No Data Connect services found in firebase.json." +
-        `\nYou can run ${clc.bold("firebase init dataconnect")} to add a Data Connect service.`,
-    );
-  } else if (serviceCfgs.length === 1) {
-    serviceInfo = await load(projectId, config, serviceCfgs[0].source);
-    if (serviceId && serviceId !== serviceInfo.dataConnectYaml.serviceId) {
-      throw new FirebaseError(
-        `No service named ${serviceId} declared in firebase.json. Found ${serviceInfo.dataConnectYaml.serviceId}.` +
-          `\nYou can run ${clc.bold("firebase init dataconnect")} to add this Data Connect service.`,
-      );
-    }
-    return serviceInfo;
-  } else {
-    if (!serviceId) {
-      throw new FirebaseError(
-        "Multiple Data Connect services found in firebase.json. Please specify a service ID to use.",
-      );
-    }
-    const infos = await Promise.all(serviceCfgs.map((c) => load(projectId, config, c.source)));
-    // TODO: handle cases where there are services with the same ID in 2 locations.
-    const maybe = infos.find((i) => i.dataConnectYaml.serviceId === serviceId);
-    if (!maybe) {
-      throw new FirebaseError(
-        `No service named ${serviceId} declared in firebase.json. Found ${infos.map((i) => i.dataConnectYaml.serviceId).join(", ")}.` +
-          `\nYou can run ${clc.bold("firebase init dataconnect")} to add this Data Connect service.`,
-      );
-    }
-    return maybe;
-  }
-}
 
 // case insensitive exact match indicators for supported app platforms
 const WEB_INDICATORS = ["package.json", "package-lock.json", "node_modules"];
