+++ src/mcp/tools/dataconnect/execute_graphql.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 import { tool } from "../../tool";
 import * as dataplane from "../../../dataconnect/dataplaneClient";
-import { pickService } from "../../../dataconnect/fileUtils";
+import { pickService } from "../../../dataconnect/load";
 import { graphqlResponseToToolResponse, parseVariables } from "./converter";
 import { Client } from "../../../apiv2";
 import { getDataConnectEmulatorClient } from "./emulator";
